Breaking Change on Immutable v4.0.0-rc.5 (https://github.com/immutable-js/immutable-js/releases/tag/v4.0.0-rc.5) :
No longer deeply coerce argument to merge()

Previously, the argument provided to merge() was deeply converted to Immutable collections via fromJS(). This was the only function in the library which calls fromJS() indirectly directly, and it was surprising and made it difficult to understand what the result of merge() would be. Now, the value provided to merge() is only shallowly converted to an Immutable collection, similar to related methods in the library. This may change the behavior of your calls to merge().